Refactor ScheduleCab component and move styles,css inside it
