+++ src/components/ScheduleCab/styles.css
@@ -1,15 +1,3 @@
-@import '@radix-ui/colors/blackA.css';
-@import '@radix-ui/colors/green.css';
-@import '@radix-ui/colors/mauve.css';
-@import '@radix-ui/colors/violet.css';
-
-/* reset */
-button,
-fieldset,
-input {
-  all: unset;
-}
-
 .DialogOverlay {
   background-color: var(--blackA9);
   position: fixed;
@@ -18,7 +6,6 @@ input {
 }
 
 .DialogContent {
-  background-color: white;
   border-radius: 6px;
   box-shadow: hsl(206 22% 7% / 35%) 0px 10px 38px -10px, hsl(206 22% 7% / 20%) 0px 10px 20px -15px;
   position: fixed;
@@ -30,16 +17,20 @@ input {
   max-height: 85vh;
   padding: 25px;
   animation: contentShow 150ms cubic-bezier(0.16, 1, 0.3, 1);
+  background-color: white;
+;
 }
 .DialogContent:focus {
-  outline: none;
+  outline: #067A6F;
 }
 
 .DialogTitle {
+  text-align: center;
   margin: 0;
   font-weight: 500;
   color: var(--mauve12);
   font-size: 17px;
+  padding: 0 0 0 0;
 }
 
 .DialogDescription {
@@ -61,6 +52,7 @@ input {
   height: 35px;
 }
 .Button.violet {
+  text-transform: revert;
   background-color: white;
   color: var(--violet11);
   box-shadow: 0 2px 10px var(--blackA7);
@@ -72,6 +64,7 @@ input {
   box-shadow: 0 0 0 2px black;
 }
 .Button.green {
+  text-transform: revert;
   background-color: var(--green4);
   color: var(--green11);
 }
@@ -82,28 +75,29 @@ input {
   box-shadow: 0 0 0 2px var(--green7);
 }
 .Button.save {
-  width: 25%;
-  padding-top: 15px;
-  padding-bottom: 15px;
-  background-color: var(--violet11);
+  text-transform: revert;
+  width: 30%;
+  padding: 15px 0;
+  background-color: #0E9888;
   color: white;
 }
 .Button.save:hover {
-  background-color: var(--violet11);
+  background-color: #53B9AB;
+  border: 2px solid #067A6F;
 }
 
 .IconButton {
   font-family: inherit;
   border-radius: 100%;
-  height: 25px;
-  width: 25px;
+  height: 20px;
+  width: 20px;
   display: inline-flex;
   align-items: center;
   justify-content: center;
   color: var(--violet11);
   position: absolute;
-  top: 10px;
-  right: 10px;
+  top: 5px;
+  right: 5px;
 }
 .IconButton:hover {
   background-color: var(--violet4);
@@ -125,8 +119,18 @@ input {
   width: 90px;
   text-align: right;
 }
+
+.RadioGroup {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  width: 125%;
+  padding: 0 10px 15px 10px;
+}
+
 .Picker {
   width: 75%;
+  height: 50%;
 }
 
 .Input {
@@ -149,22 +153,3 @@ input {
   box-shadow: 0 0 0 2px var(--violet8);
 }
 
-@keyframes overlayShow {
-  from {
-    opacity: 0;
-  }
-  to {
-    opacity: 1;
-  }
-}
-
-@keyframes contentShow {
-  from {
-    opacity: 0;
-    transform: translate(-50%, -48%) scale(0.96);
-  }
-  to {
-    opacity: 1;
-    transform: translate(-50%, -50%) scale(1);
-  }
-}
