+++ src/components/services/url.jsx
@@ -1,3 +1,3 @@
 export const API_URL = {
-   addAddressDetails : "https://cab-schedule-serverless.vercel.app/api/v1/cab-request"
+  addAddressDetails: 'https://cab-schedule-serverless.vercel.app/api/v1/cab-request',
 };
